requirements.txt
numpy
pandas
Flask
Flask-Cors
pymongo
faiss-cpu
scikit-learn
requests
packaging
python-dotenv
chromadb
langchain
langchain-core
langchain-community
langchain-openai
umap-learn
